feature/KOJO-36 | throw a TypeError some times

+++ Function42/src/V1/Worker/Delegate.php
@@ -9,9 +9,13 @@ class Delegate implements DelegateInterface
 {
     use Worker\Queue\Message\AwareTrait;
 
-    public function businessLogic(): DelegateInterface
+    public function businessLogic() : DelegateInterface
     {
-        $this->getV1WorkerQueueMessage()->delete();
+        if ((bool)random_int(0, 1)) {
+            $this->getV1WorkerQueueMessage()->delete();
+        } else {
+            throw new \TypeError("Didn't get lucky");
+        }
 
         return $this;
     }
